+++ tests/testthat/test-tags.r
@@ -399,6 +399,25 @@ test_that("Adding attributes to tags", {
   )
 })
 
+test_that("Adding unnamed attributes creates a warning", {
+  expect_error(
+    tagAppendAttributes(
+      tags$div(),
+      "value"
+    ),
+    "include an attribute name"
+  )
+
+  x <- div()
+  x$attribs[[1]] <- "value"
+  expect_error(
+    print(x),
+    "name all of your attribute values"
+  )
+})
+
+
+
 test_that("Testing for attributes on tags", {
   t1 <- tags$div("foo", class = "c1", class = "c2", id = "foo")
 
@@ -894,6 +913,7 @@ test_that("extractPreserveChunks works for emoji strings", {
   )
 })
 
+
 test_that("complicated class attributes are handled", {
   x <- div(class = as.factor(letters)[1], class = "b c", class = c("d", "e f"))
   expect_equal(
@@ -904,5 +924,70 @@ test_that("complicated class attributes are handled", {
     as.character(x),
     "<div class=\"a b c d e f\"></div>"
   )
+})
 
+
+test_that("html render method", {
+  local_edition(3)
+
+  # Have a place holder div and return a span instead
+  obj <- div("example", .renderHook = function(x) {
+    x$name <- "span"
+    x
+  })
+  expect_equal(obj$name, "div")
+  expect_snapshot(as.character(obj))
+
+  # Add a class to the tag
+  spanExtra <- tagAddRenderHook(obj, function(x) {
+    tagAppendAttributes(x, class = "extra")
+  })
+  expect_equal(spanExtra$name, "div")
+  expect_equal(spanExtra$attribs$class, NULL)
+  expect_snapshot(as.character(spanExtra))
+
+  # Replace the previous render method
+  # Should print a `div` with class `"extra"`
+  divExtra <- tagAddRenderHook(obj, replace = TRUE, function(x) {
+    tagAppendAttributes(x, class = "extra")
+  })
+  expect_equal(divExtra$attribs$class, NULL)
+  expect_snapshot(as.character(divExtra))
+
+  # Add more child tags
+  spanExtended <- tagAddRenderHook(obj, function(x) {
+    tagAppendChildren(x, tags$strong("bold text"))
+  })
+  expect_equal(spanExtended$name, "div")
+  expect_equal(spanExtended$children, obj$children)
+  expect_snapshot(as.character(spanExtended))
+
+  tagFuncExt <- tagAddRenderHook(obj, function(x) {
+    tagFunction(function() tagList(x, tags$p("test")) )
+  })
+  expect_equal(tagFuncExt$name, "div")
+  expect_equal(tagFuncExt$children, obj$children)
+  expect_snapshot(as.character(tagFuncExt))
+
+  # Add a new html dependency
+  newDep <- tagAddRenderHook(obj, function(x) {
+    fa <- htmlDependency(
+      "font-awesome", "4.5.0", c(href="shared/font-awesome"),
+      stylesheet = "css/font-awesome.min.css")
+    attachDependencies(x, fa, append = TRUE)
+  })
+  # Also add a jqueryui html dependency
+  htmlDependencies(newDep) <- htmlDependency(
+    "jqueryui", "1.11.4", c(href="shared/jqueryui"),
+    script = "jquery-ui.min.js")
+  expect_equal(newDep$name, "div")
+  expect_length(htmlDependencies(newDep), 1)
+  expect_snapshot(renderTags(newDep))
+
+  # Ignore the original tag and return something completely new.
+  newObj <- tagAddRenderHook(obj, function(x) {
+    tags$p("Something else")
+  })
+  expect_equal(newObj$name, "div")
+  expect_snapshot(as.character(newObj))
 })